+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Affichage du vrai chemin des logs en cas d’erreur
+
 ## [2.2.0] - 2026-07-10
 
 ### Added